% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autotuner-class.R, R/data.R
\docType{class}
\name{Autotuner-class}
\alias{Autotuner-class}
\alias{Autotuner}
\title{Autotuner}
\format{An Autotuner object}
\source{
data-raw/makeTestData.R
}
\description{
This file contains the skeleton to the Autotuner class used
through out Autoutuner.

This object is a generic object designed to run the different
functions of the ms2sweeper package. The slots represent content or data
that the package uses throughout the different functions.

Autouner object used to test functions contained within the package. The
content of this object was made using the mmetspData package.
}
\section{Slots}{

\describe{
\item{\code{time}}{- A list containing vectors of scan time points
from each sample.}

\item{\code{intensity}}{- A list containing vectors of scan intensity
points from each sample.}

\item{\code{peaks}}{- Regions within each sample identified as
peaks by slidingwindow analysis.}

\item{\code{peak_table}}{- A data.frame containing information on
each peak after further processing is done to the data.}

\item{\code{peak_difference}}{- A data.frame containing information
on how peaks are eluted differently over time.}

\item{\code{metadata}}{- A data.frame containing metadata for all
samples to be run on Autotuner.}

\item{\code{file_paths}}{- A string path that leads to the samples
to be run on Autotuner.}

\item{\code{file_col}}{- A string for the column name of the
column within the metadata that has specific sample names.}

\item{\code{factorCol}}{- A string for the column name of the column
within the metadata that has specific sample class names.}
}}

