% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBounds.R
\name{checkBounds}
\alias{checkBounds}
\title{checkBounds}
\usage{
checkBounds(mass, upper = T, currentMsFile, currentIndex,
  intensityStorage, ppmEst, scans, origBound)
}
\arguments{
\item{mass}{- Specific mass being checked against adjacent scans.}

\item{upper}{- A boolean value that tells the algorithm to check indices
greater than the entered one.  If false, it will check values less thant the
entered one.}

\item{currentMsFile}{- mzR link to one of the entered raw data files.}

\item{currentIndex}{- Numerical index indicating which scan contains
feature specific information.}

\item{intensityStorage}{- comming soon}

\item{ppmEst}{- Scalar numerical value meant to represent the ppm of the
instrument.}

\item{scans}{- Set of all possible ms1 scans for the sample.}

\item{origBound}{- The original scan bound location of the peak.}
}
\value{
This function returns the last index the feature is detected.
}
\description{
Recursive function used to find how far a binned feature might
extend beyond the boundary of the originally defined TIC peak.
}
