% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEICPeaks.R
\name{checkEICPeaks}
\alias{checkEICPeaks}
\title{checkEICPeaks}
\usage{
checkEICPeaks(
  mzDb,
  header,
  observedPeak,
  massThresh = 0.005,
  useGap,
  varExpThresh,
  returnPpmPlots,
  plotDir,
  filename
)
}
\arguments{
\item{mzDb}{A list of data.frames containing the m/z and intensity values
from each scan's mass spectra.}

\item{header}{A data.fame containing metadata on the sample like
spectra type, retention time, and scan count.}

\item{observedPeak}{A list with names 'start' and 'end' containing
scalar values representing the calculated peak boundary points}

\item{massThresh}{A generous exact mass error threshold used to estimate
PPM for features.}

\item{useGap}{Parameter carried into checkEICPeaks that tells Autotuner
whether to use the gap statustic to determine the proper number of clusters
to use during ppm parameter estimation.}

\item{varExpThresh}{Numeric value representing the variance explained
threshold to use if useGap is false.}

\item{returnPpmPlots}{Boolean value that tells R to return plots for
ppm distributions.}

\item{plotDir}{Path where to store plots.}

\item{filename}{A string containing the name of the current data file being
analyzed.}
}
\value{
This function returns a peak specific set of processign parameters.
}
\description{
This function is the outer most function used to check for
individual EIC peak specific parameters.
}
