% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEICFunctions.R
\name{countMaxima}
\alias{countMaxima}
\title{countMaxima}
\usage{
countMaxima(peakBounds, sortedAllEIC, prevRange)
}
\arguments{
\item{peakBounds}{What are the bounds of the peak.}

\item{sortedAllEIC}{Original data object.}

\item{prevRange}{Longest measured peak range.}
}
\value{
maxScans - the maximum number of scans across all peaks
}
\description{
This function is designed to determine the maximum number of
scans of the longest peak in the sample. It does this using the signal
processing function used earlier. This function is used during the estimation
of the maximum peakwdith.
}
