% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autotuner-class.R
\name{createAutotuner}
\alias{createAutotuner}
\title{createAutotuner}
\usage{
createAutotuner(data_paths, runfile, file_col, factorCol)
}
\arguments{
\item{data_paths}{- A string path pointing at data files to load in Autotuner.}

\item{runfile}{- a data.frame of sample metadata.}

\item{file_col}{- Character string of the column name of the column within
the runfile that contains sample names.}

\item{factorCol}{- Character string of the column name of the column within
the runfile that contains sample type factor.}
}
\description{
This function will create a Autotuner used to extract ms2s.
}
