% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissectScans.R
\name{dissectScans}
\alias{dissectScans}
\title{dissectScans}
\usage{
dissectScans(currentMsFile, observedPeak, sampleChrom)
}
\arguments{
\item{currentMsFile}{- This is an mzR object is linked to the current sample
being investigated.}

\item{observedPeak}{- A list with 'start' and 'stop' boundaries of the
current peak.}

\item{sampleChrom}{- This is the header file containing all the metadata
for the currently loaded sample.}
}
\description{
This function is designed to extract all MS1 scan features
observed within the bounds of the current TIC peak.
}
