% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EICParams.R
\name{EICparams}
\alias{EICparams}
\title{EICparams}
\usage{
EICparams(Autotuner, massThresh, peak_table)
}
\arguments{
\item{Autotuner}{- An Autotuner objected containing sample specific raw
data.}

\item{massThresh}{- A generous exact mass error threshold used to estimate
PPM for features.}

\item{peak_table}{- table of peak width values extracted with the function
peak_width_table.}
}
\description{
This function is designed to calculate the recommended
parameters from EIC peaks. It is the main holder function for a lot of
different ones involved in calculating EIC parameters.
}
\details{
The function CheckEICPeaks handles all the peak specific
computations.
}
