% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eicParamsEsts}
\alias{eicParamsEsts}
\title{eicParamsEsts}
\format{A data.frame representing the output of the EICparams function.}
\source{
inst/script/makeTestData.R
}
\usage{
eicParamsEsts
}
\description{
This object contains a data.frame of parameter estimates generated by
running the Autotuner algorithm of the Autotuner object.
}
\keyword{datasets}
