% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_loess.R
\name{estimate_loess}
\alias{estimate_loess}
\title{estimate_loess}
\usage{
estimate_loess(xs, max_missing, sampleClasses, threshold)
}
\arguments{
\item{xs}{An xcmsSet object following peakpicking and grouping}

\item{max_missing}{Estimate for maximum missing peaks within a group to
be considered for retention time correction.}

\item{sampleClasses}{A string vector representing the classes each sample
belogns to.}

\item{threshold}{A peak width parameter to estimate retention time
correction does not overfit the data.}
}
\value{
This function will return the maximum span that will still
provide retention time correction across features over time scales
as large as the estimated peak dispersion parameters.
}
\description{
estimate_loess
}
