% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_threshold.R
\name{estimate_threshold}
\alias{estimate_threshold}
\title{estimate_threshold}
\usage{
estimate_threshold(distance_vector)
}
\arguments{
\item{distance_vector}{- This vector contains the retention time differences
between individual peaks identified within a single sample TIC.}
}
\value{
A scalar value made up of the largest value predicted to come from
peaks identical peaks plus 1.
}
\description{
AS OF 2019-07-02, THIS FUNCTION IS NO LONGER PART OF THE
AUTOTUNER ALGORITHM SINCE IT IS NO LONGER CONSIDERED TO BE NECESSARY.
At a high level, this function is used to estimate appropriate
thresholds to required to distinguish between different peaks accross
samples, and replicated ones that only vary due to retention time deviation.
It assumes that duplicate peaks are a relatively short distance from one
another relative to those that distinct.
}
\details{
To estimate threshold, all the measured distances are clustered
using k-means clustering with an itteritatively increasing number of
clusters. The itteration is complete once there are as many clusters as
peaks or when the median of the smallest distance cluster is larger than the
difference between the minimum and maximum distance values within that
cluster.
}
