% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEICFunctions.R
\name{estimatePPM}
\alias{estimatePPM}
\title{estimatePPM}
\usage{
estimatePPM(first, second)
}
\arguments{
\item{first}{Numeric of length 1 respresenting the first mass entered}

\item{second}{Numeric of length 1 representing second mass entered}
}
\value{
The ppm error between the first and second entered mass
}
\description{
This function provides a convenient way to measure ppm between
two exact masses.
}
