% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEICFunctions.R
\name{estimateSNThresh}
\alias{estimateSNThresh}
\title{estimateSNThresh}
\usage{
estimateSNThresh(no_match, sortedAllEIC, approvedPeaks)
}
\arguments{
\item{no_match}{- This is a vector of numerical indicies within the raw data
mapping to scan data considered to come from noise.}

\item{sortedAllEIC}{- This is the raw data from a single TIC peak.}

\item{approvedPeaks}{- This is a data.frame that contains information on
which peaks come from TIC data.}
}
\value{
returns an estimated s/n threshold value
}
\description{
This function is responsible for computing an estimated s/n
threshold.
}
