% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportPeaksModule.R
\name{exportPeaks}
\alias{exportPeaks}
\title{exportPeaks}
\usage{
exportPeaks(input, output, session, filledPeaks)
}
\arguments{
\item{input}{- List of UI entered values.}

\item{output}{- Values exported from server to the UI after computation.}

\item{session}{- String used to match the namespace of the UI and server}

\item{filledPeaks}{- The xcms object to get exported as a feature table.}
}
\description{
This function contains the server component of the export Peak
portion of the shiny app used within Autotuner. This is the part where a
spreadsheet is saved as output.
}
