% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportPeaksModule.R
\name{exportPeaksUI}
\alias{exportPeaksUI}
\title{exportPeaksUI}
\usage{
exportPeaksUI(id)
}
\arguments{
\item{id}{- character string used to map UI and server within App}
}
\description{
This function contains the UI component of the export Peak
portion of the shiny app used within Autotuner. This is the part where a
spreadsheet is saved as output.
}
