% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_peaks.R
\name{extract_peaks}
\alias{extract_peaks}
\title{extract_peaks}
\usage{
extract_peaks(Autotuner, returned_peaks = 10, signals)
}
\arguments{
\item{Autotuner}{- An Autotuner objected containing sample specific raw
data.}

\item{returned_peaks}{- A scalar number of peaks to return for visual
inspection. Five is the minimum possible value.
is the standard.}

\item{signals}{- List containing traces and locations where signals are
detected across all samples being checked by the algorithm.}
}
\value{
peak_table_list - a list of data.frame tables containing information
on where each where peaks are located within each sample.
}
\description{
This function is designed to extract peaks observed within the
TIC from each sample, and return their indicies for further processing.
}
