% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEICFunctions.R
\name{filterPeaksfromNoise}
\alias{filterPeaksfromNoise}
\title{filterPeaksfromNoise}
\usage{
filterPeaksfromNoise(matchedMasses)
}
\arguments{
\item{matchedMasses}{A data.frame containing information on the
retained bins.}
}
\value{
A list with entries for noise peaks and true peaks.
}
\description{
This function is desgined to perform the binning of potential
features. At this point in the algorithm, a potential feature is 2+ m/z
values that are within the generous exact mass error windown provided by
the user.
}
