% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEICFunctions.R
\name{filterPpmError}
\alias{filterPpmError}
\title{filterPpmError}
\usage{
filterPpmError(approvedPeaks)
}
\arguments{
\item{approvedPeaks}{- This is a data.frame with information on bins retained
after filtering with user input mz error threshold and continuity checks.}
}
\value{
This function returns a scalar value representing ppm error estimate.
}
\description{
This function computes an estimate for the ppm error threshold.
}
\details{
A distribution is created from the set of all ppm values identified.
The most dense peak of this distribution is assumed to represent the standard
ppm error of the data.
}
