% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_noise.R
\name{find_noise}
\alias{find_noise}
\title{find_noise}
\usage{
find_noise(time_trace)
}
\arguments{
\item{time_trace}{- chromatographic intensity trace}
}
\value{
This function determines the indexes within a chromatographic
intensity trace that are not noise. It does this by asking if a given point
is greater than the average of that point and the two bounding points on
either side of it.
}
\description{
find_noise
}
