% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTruePeaks.R
\name{findTruePeaks}
\alias{findTruePeaks}
\title{findTruePeaks}
\usage{
findTruePeaks(truePeaks, sortedAllEIC)
}
\arguments{
\item{truePeaks}{- A lsit containing indicies representing each bin.}

\item{sortedAllEIC}{- All the raw ms1 data extracted from the EIC peak.}
}
\description{
This function is designed to filter out bins that don't come from
continuous scans. The idea is that after this stage, the data is ready for
parameter estimation.
}
