% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fineParamsModule.R
\name{fineParams}
\alias{fineParams}
\title{fineParams}
\usage{
fineParams(input, output, session, params, runfile, Autotuner)
}
\arguments{
\item{input}{- List of UI entered values.}

\item{output}{- Values exported from server to the UI after computation.}

\item{session}{- String used to match the namespace of the UI and server}

\item{params}{- The reactive object containing all calculated parameter
estimates from the data.}

\item{runfile}{- A data.frame containing metadata for all the samples within
the selected folder.}

\item{Autotuner}{- Data with specifics on samples path and metadata.}
}
\description{
This function contains the server component of the EIC based
parameter selection portion of the shiny app used within Autotuner. This
module is specifically involved in presentation of recommended EIC peaks
following the completion of Autotuner. It can help the user export a
the calculated parameters as a spreadsheet.
}
