% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fineParamsModule.R
\name{fineParamsUI}
\alias{fineParamsUI}
\title{fineParamsUI}
\usage{
fineParamsUI(id)
}
\arguments{
\item{id}{- character string used to map UI and server within App}
}
\description{
This function contains the UI component of the EIC based
parameter selection portion of the shiny app used within Autotuner.
}
