% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autotuner-class.R
\name{getAutoFactorCol}
\alias{getAutoFactorCol}
\title{getAutoFactorCol}
\usage{
getAutoFactorCol(Autotuner)
}
\arguments{
\item{Autotuner}{An AutoTuner object.}
}
\value{
The content of the factorCol slot
}
\description{
This function returns the character string stored within
the 'factorCol' slot of the Autotuner Object
}
\examples{
Autotuner <- readRDS(system.file("extdata/Autotuner.rds",
package="Autotuner"))
intensity <- getAutoFactorCol(Autotuner)
}
