% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autotuner-class.R
\name{getAutoFile_col}
\alias{getAutoFile_col}
\title{getAutoFile_col}
\usage{
getAutoFile_col(Autotuner)
}
\arguments{
\item{Autotuner}{An AutoTuner object.}
}
\value{
The content of the file_col slot
}
\description{
This function returns the character string stored within
the 'file_col' slot of the Autotuner Object
}
\examples{
Autotuner <- readRDS(system.file("extdata/Autotuner.rds",
package="Autotuner"))
intensity <- getAutoFile_col(Autotuner)
}
