% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autotuner-class.R
\name{getAutoFile_paths}
\alias{getAutoFile_paths}
\title{getAutoFile_paths}
\usage{
getAutoFile_paths(Autotuner)
}
\arguments{
\item{Autotuner}{An AutoTuner object.}
}
\value{
The content of the file_paths slot
}
\description{
This function returns the character string stored within
the 'file_paths' slot of the Autotuner Object
}
\examples{
Autotuner <- readRDS(system.file("extdata/Autotuner.rds",
package="Autotuner"))
intensity <- getAutoFile_paths(Autotuner)
}
