% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autotuner-class.R
\name{getAutoIntensity}
\alias{getAutoIntensity}
\title{getAutoIntensity}
\usage{
getAutoIntensity(Autotuner)
}
\arguments{
\item{Autotuner}{- An AutoTuner object.}
}
\value{
The content of the intensity slot
}
\description{
This function is designed to return the list intensities
obtained from applying the sliding window analysis to the raw data stored
within an AutoTuner object.
}
\examples{
Autotuner <- readRDS(system.file("extdata/Autotuner.rds",
package="Autotuner"))
getAutoIntensity(Autotuner)
}
