% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autotuner-class.R
\name{getAutoMetadata}
\alias{getAutoMetadata}
\title{getAutoMetadata}
\usage{
getAutoMetadata(Autotuner)
}
\arguments{
\item{Autotuner}{An AutoTuner object.}
}
\value{
The content of the meatadata slot
}
\description{
This function returns the data.frame stored within
the 'meatadata' slot of the Autotuner Object
}
\examples{
Autotuner <- readRDS(system.file("extdata/Autotuner.rds",
package="Autotuner"))
intensity <- getAutoMetadata(Autotuner)
}
