% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autotuner-class.R
\name{getAutoPeak_table}
\alias{getAutoPeak_table}
\title{getAutoPeak_table}
\usage{
getAutoPeak_table(Autotuner)
}
\arguments{
\item{Autotuner}{An AutoTuner object.}
}
\value{
The content of the peak_table slot
}
\description{
This function returns the data.frame stored within
the 'peak_table' slot of the Autotuner Object
}
\examples{
Autotuner <- readRDS(system.file("extdata/Autotuner.rds",
package="Autotuner"))
intensity <- getAutoPeak_table(Autotuner)
}
