% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autotuner-class.R
\name{getAutoPeaks}
\alias{getAutoPeaks}
\title{getAutoPeaks}
\usage{
getAutoPeaks(Autotuner)
}
\arguments{
\item{Autotuner}{An AutoTuner object.}
}
\value{
The content of the peaks slot
}
\description{
This function returns the list of numerics stored within
the 'peaks' slot of the Autotuner Object
}
\examples{
Autotuner <- readRDS(system.file("extdata/Autotuner.rds",
package="Autotuner"))
intensity <- getAutoPeaks(Autotuner)
}
