% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{header}
\alias{header}
\title{header}
\format{A data frame with 926 rows and 30 variables}
\source{
data-raw/makeAutotunerTestData.R
}
\usage{
header(object, scans, ...)
}
\description{
A feature table containing metadata from one of the mmetspData package
files used for examples, tests, and vignettes in this package.
}
