% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autotuner-class.R
\name{initialize,Autotuner-method}
\alias{initialize,Autotuner-method}
\title{Update an \code{\linkS4class{Autotuner}} object}
\usage{
\S4method{initialize}{Autotuner}(.Object, data_paths, runfile, file_col, factorCol)
}
\arguments{
\item{.Object}{The \code{\linkS4class{Autotuner}} object to update.}

\item{data_paths}{A string path pointing at data files to
load in Autotuner.}

\item{runfile}{A data.frame of sample metadata.}

\item{file_col}{Character string of the column name of the
column within the runfile that contains sample names.}

\item{factorCol}{Character string of the column name of the
column within the runfile that contains sample type factor.}
}
\value{
An updated \code{\linkS4class{Autotuner}} containing all data from
the input object.
}
\description{
This method updates an \code{\linkS4class{Autotuner}}
    object to the latest definition.
}
\author{
Craig McLean
}
