% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolatePeaks.R
\name{isolatePeaks}
\alias{isolatePeaks}
\title{isolatePeaks}
\usage{
isolatePeaks(Autotuner, returned_peaks, signals)
}
\arguments{
\item{Autotuner}{- Autotuner object. Ideally generated right after signal
processing peak identification is complete.}

\item{returned_peaks}{- numerical value representing the number of peaks
that are to be returned to the user for downstream parameter optimization.}

\item{signals}{- List of list containing data from sliding window analysis.}
}
\value{
Returns an Autotuner object with selected TIC regions.
}
\description{
This function is designed to handle the isolation of TIC peak
regions to throw into AutoTuner.
}
