% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataModule.R
\name{loadData}
\alias{loadData}
\title{loadData}
\usage{
loadData(input, output, session)
}
\arguments{
\item{input}{- List of UI entered values.}

\item{output}{- Values exported from server to the UI after computation.}

\item{session}{- String used to match the namespace of the UI and server}
}
\description{
This function contains the server component of the data loading
portion of the shiny app used within Autotuner.
}
