% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataModule.R
\name{loadDataUI}
\alias{loadDataUI}
\title{loadDataUI}
\usage{
loadDataUI(id)
}
\arguments{
\item{id}{- character string used to map UI and server within App'}
}
\description{
This function contains the UI component of the data loading
portion of the shiny app used within Autotuner.
}
