% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mzDb}
\alias{mzDb}
\title{mzDb}
\format{
A list with length 926
}
\source{
inst/script/makeAutotunerTestData.R
}
\usage{
mzDb
}
\description{
A list containing all scan information from one of the mmetspData package
files used for examples, tests, and vignettes in this package.
}
\keyword{datasets}
