% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{observedPeak}
\alias{observedPeak}
\title{observedPeak}
\format{A list with length 2
\describe{
  \item{start}{56.18}
  \item{end}{97.73}
}}
\source{
data-raw/makeAutotunerTestData.R
}
\usage{
observedPeak
}
\description{
A list containing peak start and end times for one peak of the mmetspData
package files used for examples, tests, and vignettes in this package.
}
\keyword{datasets}
