% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peak_difference}
\alias{peak_difference}
\title{peak_difference}
\format{A data frame with 7 rows and 9 columns}
\source{
inst/script/makeTestData.R
}
\usage{
peak_difference
}
\description{
A data.frame generated by during the peak identification phase of Autotuner
package files used for examples, tests, and vignettes in this package.
}
\keyword{datasets}
