% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peak_table}
\alias{peak_table}
\title{peak_table}
\format{A data frame with 30 rows and 9 columns}
\source{
data-raw/makeTestData.R
}
\usage{
peak_table
}
\description{
A data.frame generated by during the peak identification phase of Autotuner
package files used for examples, tests, and vignettes in this package.
}
\keyword{datasets}
