% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_time_difference.R
\name{peak_time_difference}
\alias{peak_time_difference}
\title{peak_time_difference}
\usage{
peak_time_difference(peak_table)
}
\arguments{
\item{peak_table}{- table of peak width values extracted with the function
peak_width_table.}
}
\value{
This function returns a data.frame of peaks matched over time.
}
\description{
This function is designed to return a data.frame containing info
on how
}
\details{
This function is designed to determine what are the retention time
differences between peaks that are effectively the same between samples.
The similarity in peaks is determined by a threshold in retention time
similarity between peaks. This function returns the max peak width between
samples, and the time difference between peaks across samples in a data frame
object. The current and next row indexes are given to go back to the
peaktable object to plot peaks.
}
