% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakVisModule.R
\name{peakVis}
\alias{peakVis}
\title{peakVis}
\usage{
peakVis(input, output, session, signalData, Autotuner)
}
\arguments{
\item{input}{- List of UI entered values.}

\item{output}{- Values exported from server to the UI after computation.}

\item{session}{- String used to match the namespace of the UI and server}

\item{signalData}{- Information on peaks selected from sliding windown
analysis in the previous page of shiny app.}

\item{Autotuner}{- Data with specifics on samples path and metadata.}
}
\description{
This function contains the server component of the peak
visualization portion of the shiny app used within Autotuner. In this section
the user selects if signal processing parameters were properly tuned to
indentify individual TIC peaks.
}
