% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakVisModule.R
\name{peakVisUI}
\alias{peakVisUI}
\title{@title peakVisUI}
\usage{
peakVisUI(id)
}
\arguments{
\item{id}{- character string used to map UI and server within App}
}
\description{
@description This file contains all the code needed to run the peak
 visualization from the metabolomics data analysis shiny page. The main
 functions for constructing UI and server behavior are teh "peakVisUI" and
 "peakVis" respectively. All of the other functions depend on
}
