% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakwidth_est.R
\name{peakwidth_est}
\alias{peakwidth_est}
\title{peakwidth_est}
\usage{
peakwidth_est(
  peak_vector,
  time,
  intensity,
  start = NULL,
  end = NULL,
  old_r2 = NULL
)
}
\arguments{
\item{peak_vector}{A numeric vector with names of specific time points of the
chromatography data measured. The numeric values correspond to indiciess
within the total chromatographic data that span the peak width.}

\item{time}{This vector contains the time measurements during the
chromatography. This vector is used to match the values in peak_vector
to the names in the intensity vector.}

\item{intensity}{A measured intensity values for chromatorgraphy}

\item{start}{A numeric index indicating where peak starts. Leave null.}

\item{end}{The same as above, leave null.}

\item{old_r2}{A previous fit of model used to judge recursion of fit.}
}
\value{
This function returns a scalar value representing the estimated peak
width for a given peak.
}
\description{
This function is designed to generate peak width estimates for
each TIC peak detected by sliding window analysis.
}
\details{
This function takes in one peak vector at a time and runs a linear
model on the selected start and end points of a peak. By measuring the
change of the fit of the model, the function returns an index of values
corresponding to a peak. This function works recursively to estimate
the width of the peak. Ultimately, it returns the names of the final
points in the peak.
}
