% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakwidth_table.R
\name{peakwidth_table}
\alias{peakwidth_table}
\title{peakwidth_table}
\usage{
peakwidth_table(Autotuner, returned_peaks = 10)
}
\arguments{
\item{Autotuner}{An Autotuner objected containing sample specific raw
data.}

\item{returned_peaks}{A scalar number of peaks to return for visual
inspection. Five is the minimum possible value.}
}
\value{
This function will return a peak table with information on
the peak width for each detected peak across samples, the name
attribute for when the peak starts and ends, and the time points
associated with each of those parameters and for the midpoint.
}
\description{
This function is designed to generate estimates of peakwidth
for each peak within the peakList and some properties of each peak. After
this is done, the table of estimates is exported.
}
\details{
The actual calculations used to estimate peakwidth are done within
the function "peakwidth_est".
}
