% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peaks.R
\name{plot_peaks}
\alias{plot_peaks}
\title{@title plot_peaks}
\usage{
plot_peaks(Autotuner, boundary = 10, peak)
}
\arguments{
\item{Autotuner}{An Autotuner objected containing sample specific raw
data.}

\item{boundary}{UI input value that defines the boundary around the peak to
visualize it.}

\item{peak}{A Numeric index obtained from UI that indicates which peak
should be visualized.}
}
\value{
This function outputs plots that are meant to go into the peakVis
UI.
}
\description{
@description This funciton plots the peak identified within
 chromatography.
}
\details{
This function plots individual peaks selected by signal
processing and expanded with a regression to allow the user to validate the
selected signal processing parameters.
}
\examples{
Autotuner <- readRDS(system.file("extdata/Autotuner.rds",
package="Autotuner"))
plot_peaks(Autotuner = Autotuner, boundary = 100, peak = 1)

}
