% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnParams.R
\name{returnParams}
\alias{returnParams}
\title{returnParams}
\usage{
returnParams(eicParamEsts, Autotuner)
}
\arguments{
\item{eicParamEsts}{- The objection containing all parameter estimates
obtained from running Autotuner's EICparam function.}

\item{Autotuner}{- An Autotuner object used to return the TIC estimated
parameters}
}
\value{
A list of data.frames with parameter estimates.
}
\description{
This function is designed to return a list of data.frames
containing parameter estimates obtained from the EIC and TIC parameter
estimation.
}
\examples{
Autotuner <- readRDS(system.file("extdata/Autotuner.rds",
package="Autotuner"))
EICparams(Autotuner = Autotuner, massThresh = .005, verbose = F,
returnPpmPlots = F, useGap = T)
returnParams(eicParamEsts, Autotuner)

}
