% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autotuner-class.R
\name{setAutoFactorCol}
\alias{setAutoFactorCol}
\title{setAutoFactorCol}
\usage{
setAutoFactorCol(factorCol, Autotuner)
}
\arguments{
\item{factorCol}{A character vector representing factorCol}

\item{Autotuner}{An AutoTuner object.}
}
\value{
An Autotuner object with a filled factorCol slot
}
\description{
This function fills the factorCol slot within an Autotuner
object.
}
\examples{
Autotuner <- readRDS(system.file("extdata/Autotuner.rds",
package="Autotuner"))
factorCol <- getAutoFactorCol(Autotuner)
Autotuner <- setAutoFactorCol(factorCol, Autotuner)
}
