% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autotuner-class.R
\name{setAutoFile_col}
\alias{setAutoFile_col}
\title{setAutoFile_col}
\usage{
setAutoFile_col(file_col, Autotuner)
}
\arguments{
\item{file_col}{A character vector representing file_col}

\item{Autotuner}{An AutoTuner object.}
}
\value{
An Autotuner object with a filled file_col slot
}
\description{
This function fills the file_col slot within an Autotuner
object.
}
\examples{
Autotuner <- readRDS(system.file("extdata/Autotuner.rds",
package="Autotuner"))
file_col <- getAutoFile_col(Autotuner)
Autotuner <- setAutoFile_col(file_col, Autotuner)
}
