% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autotuner-class.R
\name{setAutoFile_paths}
\alias{setAutoFile_paths}
\title{setAutoFile_paths}
\usage{
setAutoFile_paths(file_paths, Autotuner)
}
\arguments{
\item{file_paths}{A character vector representing file_paths}

\item{Autotuner}{An AutoTuner object.}
}
\value{
An Autotuner object with a filled file_paths slot
}
\description{
This function fills the file_paths slot within an Autotuner
object.
}
\examples{
Autotuner <- readRDS(system.file("extdata/Autotuner.rds",
package="Autotuner"))
file_paths <- getAutoFile_paths(Autotuner)
Autotuner <- setAutoFile_paths(file_paths = file_paths, Autotuner)
}
