% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autotuner-class.R
\name{setAutoIntensity}
\alias{setAutoIntensity}
\title{setAutoIntensity}
\usage{
setAutoIntensity(intensity, Autotuner)
}
\arguments{
\item{intensity}{A list of numeric values representing intensity}

\item{Autotuner}{An AutoTuner object.}
}
\value{
An Autotuner object with a filled intensity slot
}
\description{
This function fills the "intensity" slot within an Autotuner
object.
}
\examples{
Autotuner <- readRDS(system.file("extdata/Autotuner.rds",
package="Autotuner"))
intensity <- getAutoIntensity(Autotuner)
Autotuner <- setAutoIntensity(intensity, Autotuner)
}
