% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autotuner-class.R
\name{setAutoMetadata}
\alias{setAutoMetadata}
\title{setAutoMetadata}
\usage{
setAutoMetadata(metadata, Autotuner)
}
\arguments{
\item{metadata}{A data.frame representing metadata}

\item{Autotuner}{An AutoTuner object.}
}
\value{
An Autotuner object with a filled metadata slot
}
\description{
This function fills the metadata slot within an Autotuner
object.
}
\examples{
Autotuner <- readRDS(system.file("extdata/Autotuner.rds",
package="Autotuner"))
metadata <- getAutoMetadata(Autotuner)
Autotuner <- setAutoMetadata(metadata, Autotuner)
}
