% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autotuner-class.R
\name{setAutoPeak_difference}
\alias{setAutoPeak_difference}
\title{setAutoPeak_difference}
\usage{
setAutoPeak_difference(peak_difference, Autotuner)
}
\arguments{
\item{peak_difference}{A data.frame representing peak_difference}

\item{Autotuner}{An AutoTuner object.}
}
\value{
An Autotuner object with a filled peak_difference slot
}
\description{
This function fills the peak_difference slot within an
Autotuner object.
}
\examples{
Autotuner <- readRDS(system.file("extdata/Autotuner.rds",
package="Autotuner"))
peak_difference <- getAutoPeak_table(Autotuner)
Autotuner <- setAutoPeak_difference(peak_difference, Autotuner)
}
