% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autotuner-class.R
\name{setAutoPeak_table}
\alias{setAutoPeak_table}
\title{setAutoPeak_table}
\usage{
setAutoPeak_table(peak_table, Autotuner)
}
\arguments{
\item{peak_table}{A data.frame representing peak_table}

\item{Autotuner}{An AutoTuner object.}
}
\value{
An Autotuner object with a filled peak_table slot
}
\description{
This function fills the peak_table slot within an Autotuner
object.
}
\examples{
Autotuner <- readRDS(system.file("extdata/Autotuner.rds",
package="Autotuner"))
peak_table <- getAutoPeak_table(Autotuner)
Autotuner <- setAutoPeak_table(peak_table, Autotuner)
}
