% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autotuner-class.R
\name{setAutoPeaks}
\alias{setAutoPeaks}
\title{setAutoPeaks}
\usage{
setAutoPeaks(peaks, Autotuner)
}
\arguments{
\item{peaks}{A list of numeric values representing peaks}

\item{Autotuner}{An AutoTuner object.}
}
\value{
An Autotuner object with a filled peaks slot
}
\description{
This function fills the peaks slot within an Autotuner object.
}
\examples{
Autotuner <- readRDS(system.file("extdata/Autotuner.rds",
package="Autotuner"))
peaks <- getAutoPeaks(Autotuner)
Autotuner <- setAutoPeaks(peaks, Autotuner)
}
