% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autotuner-class.R
\name{setAutoTime}
\alias{setAutoTime}
\title{setAutoTime}
\usage{
setAutoTime(time, Autotuner)
}
\arguments{
\item{time}{A list of numeric values representing time}

\item{Autotuner}{An AutoTuner object.}
}
\value{
An Autotuner object with a filled time slot
}
\description{
This function fills the "time" slot within an Autotuner object.
}
\examples{
Autotuner <- readRDS(system.file("extdata/Autotuner.rds",
package="Autotuner"))
time <- getAutoTime(Autotuner)
Autotuner <- setAutoTime(time, Autotuner)
}
