% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalProcessingModule.R
\name{signalProcessing}
\alias{signalProcessing}
\title{signalProcessing}
\usage{
signalProcessing(input, output, session, Autotuner)
}
\arguments{
\item{input}{- List of UI entered values.}

\item{output}{- Values exported from server to the UI after computation.}

\item{session}{- String used to match the namespace of the UI and server}

\item{Autotuner}{- An Autotuner objected containing sample specific raw
data.}
}
\description{
This function contains the server component of the signal
processing portion of the shiny app used within Autotuner.
}
