% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalProcessingModule.R
\name{signalProcessingUI}
\alias{signalProcessingUI}
\title{signalProcessingUI}
\usage{
signalProcessingUI(id)
}
\arguments{
\item{id}{- character string used to map UI and server within App}
}
\description{
This function contains the UI component of the signal processing
portion of the shiny app used within Autotuner.
}
