% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIC_params.R
\name{TIC_params}
\alias{TIC_params}
\title{TIC_params}
\usage{
TIC_params(peak_table, peak_difference)
}
\arguments{
\item{peak_table}{- A data.frame containing information on peak width values
extracted with the function peak_width_table.}

\item{peak_difference}{- A data.frame containing information on retention
time differences between peaks.}
}
\value{
Returns a set of parameters to run xcms.
}
\description{
This function is designed to return the parameters related to
chromatography gathered during the TIC peak selection steps. An estimate is
given for maximum and minimum peak width as well as the bandwidth parameter
used in grouping.
}
