% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autotuner-class.R
\name{createAutotuner}
\alias{createAutotuner}
\title{createAutotuner}
\usage{
createAutotuner(data_paths, runfile, file_col, factorCol)
}
\arguments{
\item{data_paths}{- A string path pointing at data files to load in Autotuner.}

\item{runfile}{- a data.frame of sample metadata.}

\item{file_col}{- Character string of the column name of the column within
the runfile that contains sample names.}

\item{factorCol}{- Character string of the column name of the column within
the runfile that contains sample type factor.}
}
\description{
This function will create a Autotuner used to extract ms2s.
}
\examples{
library(devtools)
if(!require("mmetspData")) {
    install_github("crmclean/mmetspData")
}
library(mmetspData)
mmetspFiles <- c(system.file("mzMLs/mtab_mmetsp_ft_120815_24.mzML",
package = "mmetspData"), system.file("mzMLs/mtab_mmetsp_ft_120815_25.mzML",
package = "mmetspData"), system.file("mzMLs/mtab_mmetsp_ft_120815_26.mzML",
package = "mmetspData"))
metadata <- read.csv(system.file("mmetsp_metadata.csv",
package = "mmetspData"),stringsAsFactors = FALSE)
metadata <- metadata[metadata$File.Name \%in\%
sub(pattern = ".mzML", "",basename(mmetspFiles)),]
Autotuner <- createAutotuner(mmetspFiles, metadata,
file_col = "File.Name", factorCol = "Sample.Type")


}
