% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissectScans.R
\name{dissectScans}
\alias{dissectScans}
\title{dissectScans}
\usage{
dissectScans(mzDb, observedPeak, header)
}
\arguments{
\item{mzDb}{- This is a list of two column data frames containing information
on each mass spectra within the data.}

\item{observedPeak}{- A list with 'start' and 'stop' boundaries of the
current peak.}

\item{header}{- This is the header file containing all the metadata
for the currently loaded sample.}
}
\description{
This function is designed to extract all MS1 scan features
observed within the bounds of the current TIC peak.
}
