% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTruePeaks.R
\name{findTruePeaks}
\alias{findTruePeaks}
\title{findTruePeaks}
\usage{
findTruePeaks(truePeaks, sortedAllEIC)
}
\arguments{
\item{truePeaks}{A list containing indicies representing each bin.}

\item{sortedAllEIC}{All the raw ms1 data extracted from the EIC peak.}
}
\value{
a list of candidate EIC regions
}
\description{
This function is designed to filter out bins that don't come
from continuous scans. The idea is that after this stage, the data is ready
for parameter estimation.
}
