% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{header}
\alias{header}
\title{header}
\format{A data frame with 926 rows and 30 variables:
\describe{
  \item{spectrumId}{Speciic spectrum collected by data}
  \item{spectrum}{Index for each spectrum}
  ...
}}
\source{
data-raw/makeAutotunerTestData.R
}
\usage{
header
}
\description{
A feature table containing metadata from one of the mmetspData package
files used for examples, tests, and vignettes in this package.
}
\keyword{datasets}
