% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolatePeaks.R
\name{isolatePeaks}
\alias{isolatePeaks}
\title{isolatePeaks}
\usage{
isolatePeaks(Autotuner, returned_peaks, signals)
}
\arguments{
\item{Autotuner}{An Autotuner object. Ideally generated right after signal
processing peak identification is complete.}

\item{returned_peaks}{A numerical value representing the number of peaks
that are to be returned to the user for downstream parameter optimization.}

\item{signals}{A list of list containing data from sliding window analysis.}
}
\value{
Returns an Autotuner object with selected TIC regions.
}
\description{
This function is designed to handle the isolation of TIC peak
regions to throw into AutoTuner.
}
\examples{
Autotuner <- readRDS(system.file("extdata/Autotuner.rds",
package="Autotuner"))
lag <- 25
threshold<- 3.1
influence <- 0.1
signals <- lapply(getAutoIntensity(Autotuner), ThresholdingAlgo,
lag, threshold, influence)
isolatePeaks(Autotuner, returned_peaks = 10, signals)

}
