% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signals.R
\name{plot_signals}
\alias{plot_signals}
\title{plot_signals - this funciton plots the peak identified within chromatography.}
\usage{
plot_signals(Autotuner, threshold, sample_index, signals)
}
\arguments{
\item{Autotuner}{- MSest object created following Create_MSest()
initialization function.}

\item{threshold}{- User input scalar value for the number of standard
deviations required to consider a peak to be significant.}

\item{sample_index}{- which of all of the samples should the user plot.
Entered in as a numerical index value with length 1.}

\item{signals}{- A vector containing information on where signals are
located between samples.}
}
\value{
Plots signal
}
\description{
This function plots the chromatography and the matching sliding
window signal processing results for each sample. Signal processing
functions will be the same color as the chromatography spectra, just a
lighter shade and a different type of line. The chromatography will be a
solid line, the signal will be a dashed line (lty = 2) with a .5 alpha,
and the thersholds will be dotted lines (lty = 3) with alpha values of 3.
}
